+++ CHANGELOG.md
@@ -3,12 +3,12 @@
 ## UNRELEASED
 
 - Add vector retrieval and search by vector to collections [#343](https://github.com/hypermodeAI/runtime/pull/343)
+- Refactor dgraph host functions to use single execute host function [#342](https://github.com/hypermodeAI/runtime/pull/342)
 
 ## 2024-08-16 - Version 0.11.1
 
 - Improve logger registration code [#335](https://github.com/hypermodeAI/runtime/pull/335)
 - Add dgraph host functions [#336](https://github.com/hypermodeAI/runtime/pull/336)
-- Add upsert to dgraph host functions [#342](https://github.com/hypermodeAI/runtime/pull/342)
 
 ## 2024-08-12 - Version 0.11.0