+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Change Log
 
-## UNRELEASED
+## 2024-08-27 - Version 0.11.2
 
-- Add vector retrieval and search by vector to collections [#343](https://github.com/hypermodeAI/runtime/pull/343)
 - Refactor dgraph host functions to use single execute host function [#342](https://github.com/hypermodeAI/runtime/pull/342)
+- Add vector retrieval and search by vector to collections [#343](https://github.com/hypermodeAI/runtime/pull/343)
 
 ## 2024-08-16 - Version 0.11.1